pyproject.toml
[project]
name = "nexloop"
version = "1.0.0"
description = "Blueguard Marketing AI Dashboard - SOLID Architecture"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "google-genai>=0.5.0",
    "google-cloud-storage>=2.10.0",
    "google-cloud-discoveryengine>=0.11.0",
    "google-cloud-scheduler>=2.13.0",
    "google-api-python-client>=2.100.0",
    "youtube-transcript-api>=0.6.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "reportlab>=4.0.0",
    "notion-client>=2.0.0",
    "Pillow>=10.0.0",
    "PyJWT>=2.8.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt<4.0.0",
    "python-jose[cryptography]>=3.3.0",
    "alembic>=1.13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "src/api",
    "src/config",
    "src/core",
    "src/infrastructure",
    "src/services",
    "src/utils",
]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = [
    "E", "F", "I", "N", "W",
    "B",  # flake8-bugbear
    "UP", # pyupgrade
    "C4", # flake8-comprehensions
    "SIM", # flake8-simplify
    "RUF", # ruff-specific rules
    "A",   # flake8-builtins
    "ASYNC", # flake8-async
]
ignore = ["E501"]
exclude = [
    ".agent",
    ".idea",
    ".pytest_cache",
    ".ruff_cache",
    "outputs",
    "docs",
]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
strict = true
check_untyped_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
mypy_path = ["src"]
exclude = "(^tests/)"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --cov=src -p no:cacheprovider"
